Create a Message File
This challenge consist in creating a code in HTML that uses just paragraphs and line breakers to reach in a specific result.
And this is the way I made it, since I don't have no one to actually revise it. I think this is the right way to do it!